Fix warnings and clippy across workspace:\n- Clean unused imports/vars, resolve unreachable pattern in event.rs\n- Improve AI runtime init, streaming batching, div_ceil; workflow panel unwrap_or\n- Renderer minor cast cleanup; composer_edit iteration helpers\n- Add cfg(ci) registration in AI crate build.rs for clippy\n- Plugin-system: remove needless lifetimes, redundant map_err closures\n- Core: collapse redundant else blocks; native-extras cleanup\n- Completions: align CommandSpec with usage

+++ openagent-terminal/src/renderer/wgpu.rs
@@ -1351,7 +1351,7 @@ fn fs_main(in: VsOut) -> @location(0) vec4<f32> {
             // Flush staging -> GPU vertex buffer before drawing rects/UI
             let copied = self.rect_transfer.flush(&mut encoder, &self.device);
             if copied > 0 {
-                self.metrics.rect_bytes_copied = copied as u64;
+self.metrics.rect_bytes_copied = copied;
                 self.metrics.rect_flush_count += 1;
             }
